# src/ghosts.ui/src/generated/swagger.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "GHOSTS API",
		"description": "GHOSTS API v8 - Assembly: 8.0.0.0",
		"contact": {
			"name": "Dustin Updyke",
			"url": "https://sei.cmu.edu",
			"email": "ddupdyke [*at*] sei.cmu.edu"
		},
		"license": {
			"name": "Copyright 2017 Carnegie Mellon University. All Rights Reserved. See LICENSE.md file for terms"
		},
		"version": "v8"
	},
	"paths": {
...
		"/api/timelines/{machineId}": {
			"get": {
				"tags": ["TimeLines"],
				"summary": "This returns all timeLines for a requested machine. If all or a specific timeLine is not available,\r\na MachineUpdate request can be made to retrieve the machine timeLines",
				"parameters": [
					{
						"name": "machineId",
						"in": "path",
						"description": "Machine Guid",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/MachineTimeLine"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MachineTimeLine"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/MachineTimeLine"
								}
							}
						}
					}
				}
			}
		},
...
		"/api/timelines/{machineId}/{timeLineId}": {
			"get": {
				"tags": ["TimeLines"],
				"summary": "This returns a specific timeLine for a requested machine. If the timeLine is not available,\r\na MachineUpdate request can be made to retrieve the machine timeLine",
				"parameters": [
					{
						"name": "machineId",
						"in": "path",
						"description": "Machine Guid",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "timeLineId",
						"in": "path",
						"description": "TimeLine Id Guid",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/MachineTimeLine"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MachineTimeLine"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/MachineTimeLine"
								}
							}
						}
					}
				}
			}
		},
		"/api/timelines": {
			"post": {
				"tags": ["TimeLines"],
				"summary": "Send a new timeLine to a particular machine",
				"requestBody": {
					"description": "The update to send",
					"content": {
						"application/json-patch+json": {
							"schema": {
								"$ref": "#/components/schemas/MachineUpdateViewModel"
							},
							"example": {
								"machineId": "eae2268d-96ef-4a8d-9c54-e43339cb90f7",
								"type": "TimeLinePartial",
								"activeUtc": "2024-04-17T11:57:04.7536432Z",
								"status": "Active",
								"update": {
									"id": "9dd5ba70-f01b-43a9-bc1e-908b9b80964b",
									"status": "Run",
									"timeLineHandlers": [
										{
											"handlerType": "BrowserFirefox",
											"initial": "https://cmu.edu",
											"utcTimeOn": "00:00:00",
											"utcTimeOff": "23:59:59",
											"handlerArgs": {},
											"loop": true,
											"timeLineEvents": [
												{
													"command": "browse",
													"commandArgs": ["https://sei.cmu.edu"],
													"delayAfter": 30000,
													"delayBefore": 0
												}
											],
											"scheduleType": "Other"
										}
									]
								}
							}
						},
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MachineUpdateViewModel"
							},
							"example": {
								"machineId": "eae2268d-96ef-4a8d-9c54-e43339cb90f7",
								"type": "TimeLinePartial",
								"activeUtc": "2024-04-17T11:57:04.7536432Z",
								"status": "Active",
								"update": {
									"id": "9dd5ba70-f01b-43a9-bc1e-908b9b80964b",
									"status": "Run",
									"timeLineHandlers": [
										{
											"handlerType": "BrowserFirefox",
											"initial": "https://cmu.edu",
											"utcTimeOn": "00:00:00",
											"utcTimeOff": "23:59:59",
											"handlerArgs": {},
											"loop": true,
											"timeLineEvents": [
												{
													"command": "browse",
													"commandArgs": ["https://sei.cmu.edu"],
													"delayAfter": 30000,
													"delayBefore": 0
												}
											],
											"scheduleType": "Other"
										}
									]
								}
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/MachineUpdateViewModel"
							},
							"example": {
								"machineId": "eae2268d-96ef-4a8d-9c54-e43339cb90f7",
								"type": "TimeLinePartial",
								"activeUtc": "2024-04-17T11:57:04.7536432Z",
								"status": "Active",
								"update": {
									"id": "9dd5ba70-f01b-43a9-bc1e-908b9b80964b",
									"status": "Run",
									"timeLineHandlers": [
										{
											"handlerType": "BrowserFirefox",
											"initial": "https://cmu.edu",
											"utcTimeOn": "00:00:00",
											"utcTimeOff": "23:59:59",
											"handlerArgs": {},
											"loop": true,
											"timeLineEvents": [
												{
													"command": "browse",
													"commandArgs": ["https://sei.cmu.edu"],
													"delayAfter": 30000,
													"delayBefore": 0
												}
											],
											"scheduleType": "Other"
										}
									]
								}
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/MachineUpdateViewModel"
							},
							"example": {
								"machineId": "eae2268d-96ef-4a8d-9c54-e43339cb90f7",
								"type": "TimeLinePartial",
								"activeUtc": "2024-04-17T11:57:04.7536432Z",
								"status": "Active",
								"update": {
									"id": "9dd5ba70-f01b-43a9-bc1e-908b9b80964b",
									"status": "Run",
									"timeLineHandlers": [
										{
											"handlerType": "BrowserFirefox",
											"initial": "https://cmu.edu",
											"utcTimeOn": "00:00:00",
											"utcTimeOff": "23:59:59",
											"handlerArgs": {},
											"loop": true,
											"timeLineEvents": [
												{
													"command": "browse",
													"commandArgs": ["https://sei.cmu.edu"],
													"delayAfter": 30000,
													"delayBefore": 0
												}
											],
											"scheduleType": "Other"
										}
									]
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			}
		},
		"/api/timelines/{machineId}/{timeLineId}/stop": {
			"post": {
				"tags": ["TimeLines"],
				"parameters": [
					{
						"name": "machineId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "timeLineId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"HandlerType": {
				"enum": [
					"BrowserFirefox",
					"BrowserChrome",
					"Command",
					"Notepad",
					"Outlook",
					"Word",
					"Excel",
					"PowerPoint",
					"NpcSystem",
					"Reboot",
					"Curl",
					"Clicks",
					"Watcher",
					"LightWord",
					"LightExcel",
					"LightPowerPoint",
					"PowerShell",
					"Bash",
					"Print",
					"Ssh",
					"Sftp",
					"Pidgin",
					"Rdp",
					"Wmi",
					"Outlookv2",
					"Ftp",
					"AwsCli"
				],
				"type": "string"
			},
...
			"MachineTimeLine": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"machineId": {
						"type": "string",
						"format": "uuid"
					},
					"timeLine": {
						"type": "string",
						"description": "This is a string unfortunately, because EF can't handle string arrays, etc.",
						"nullable": true
					},
					"createdUtc": {
						"type": "string",
						"format": "date-time"
					}
				},
				"additionalProperties": false
			},
...
			"MachineUpdateViewModel": {
				"type": "object",
				"properties": {
					"machineId": {
						"type": "string",
						"format": "uuid"
					},
					"type": {
						"$ref": "#/components/schemas/UpdateType"
					},
					"activeUtc": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/StatusType"
					},
					"update": {
						"$ref": "#/components/schemas/TimeLine"
					}
				},
				"additionalProperties": false
			},
...
			"StatusType": {
				"enum": ["Active", "Deleted"],
				"type": "string"
			},
...
			"TimeSpan": {
				"type": "object",
				"properties": {
					"ticks": {
						"type": "integer",
						"format": "int64"
					},
					"days": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"hours": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"milliseconds": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"minutes": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"seconds": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"totalDays": {
						"type": "number",
						"format": "double",
						"readOnly": true
					},
					"totalHours": {
						"type": "number",
						"format": "double",
						"readOnly": true
					},
					"totalMilliseconds": {
						"type": "number",
						"format": "double",
						"readOnly": true
					},
					"totalMinutes": {
						"type": "number",
						"format": "double",
						"readOnly": true
					},
					"totalSeconds": {
						"type": "number",
						"format": "double",
						"readOnly": true
					}
				},
				"additionalProperties": false
			},
			"TimeLine": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"status": {
						"$ref": "#/components/schemas/TimeLineStatus"
					},
					"timeLineHandlers": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/TimeLineHandler"
						},
						"nullable": true
					}
				},
				"additionalProperties": false
			},
			"TimeLineEvent": {
				"type": "object",
				"properties": {
					"trackableId": {
						"type": "string",
						"nullable": true
					},
					"command": {
						"type": "string",
						"nullable": true
					},
					"commandArgs": {
						"type": "array",
						"items": {},
						"nullable": true
					},
					"delayAfter": {
						"nullable": true
					},
					"delayBefore": {
						"nullable": true
					},
					"delayAfterActual": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"delayBeforeActual": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					}
				},
				"additionalProperties": false
			},
			"TimeLineHandler": {
				"type": "object",
				"properties": {
					"handlerType": {
						"$ref": "#/components/schemas/HandlerType"
					},
					"initial": {
						"type": "string",
						"nullable": true
					},
					"utcTimeOn": {
						"$ref": "#/components/schemas/TimeSpan"
					},
					"utcTimeOff": {
						"$ref": "#/components/schemas/TimeSpan"
					},
					"utcTimeBlocks": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/TimeSpan"
						},
						"nullable": true
					},
					"handlerArgs": {
						"type": "object",
						"additionalProperties": {
							"nullable": true
						},
						"nullable": true
					},
					"loop": {
						"type": "boolean"
					},
					"timeLineEvents": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/TimeLineEvent"
						},
						"nullable": true
					},
					"scheduleType": {
						"$ref": "#/components/schemas/TimeLineScheduleType"
					},
					"schedule": {
						"type": "string",
						"nullable": true
					}
				},
				"additionalProperties": false
			},
			"TimeLineScheduleType": {
				"enum": ["Other", "Cron"],
				"type": "string"
			},
			"TimeLineStatus": {
				"enum": ["Run", "Stop"],
				"type": "string"
			},
...
			"UpdateType": {
				"enum": ["TimeLine", "Health", "TimeLinePartial", "RequestForTimeLine"],
				"type": "string"
			},
...
		}
	}
}